requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Twilio
twilio==8.11.0

# Langchain
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.7

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Audio processing
pydub==0.25.1
soundfile==0.12.1
numpy==1.24.3

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async support
asyncio-mqtt==0.13.0

# Logging and monitoring
loguru==0.7.2


# Development
black==23.11.0
flake8==6.1.0